Updated services

+++ src/main/java/by/bsuir/deliveryservice/dao/ShippingDao.java
@@ -4,7 +4,7 @@ import by.bsuir.deliveryservice.entity.Shipping;
 
 import java.util.List;
 
-public interface ShippingDao extends GenericDao<Integer, Shipping> {
-    Integer selectShippingIdByName(String name) throws DaoException;
+public interface ShippingDao extends GenericDao<Long, Shipping> {
+    Long selectShippingIdByName(String name) throws DaoException;
     List<Shipping> selectAllShippings() throws DaoException;
 }
