+++ src/main/java/by/bsuir/deliveryservice/service/ShippingService.java
@@ -5,6 +5,6 @@ import by.bsuir.deliveryservice.entity.Shipping;
 import java.util.List;
 
 public interface ShippingService {
-    void updateShipping(Integer shippingId, Shipping shipping) throws ServiceException;
+    void updateShipping(Long shippingId, Shipping shipping) throws ServiceException;
     List<Shipping> viewPricelist() throws ServiceException;
 }
